# openapiDoc.JSON
{
  "openapi": "3.0.3",
  "info": {
    "title": "Groupomania social network",
    "description": "This is the documentation of Groupomania social network API.",
    "termsOfService": "http://groupomania.com/terms/",
    "contact": {
      "name": "API support",
      "url": "http://groupomania.com/contact/",
      "email": "contactAPI@groupomania.com"
    },
    "license": {
      "name": "Apache 2.0",
      "url": "http://www.apache.org/licenses/LICENSE-2.0.html"
    },
    "version": "1.0.0"
  },
  "servers": [{
    "url": "http://localhost:{port}/{basePath}",
    "variables": {
      "port": {
        "default": "3001"
      },
      "basePath": {
        "default": "api"
      }
    }
  }],
...
  "paths": {
...
    "/comments": {
      "post": {
        "tags": ["comments"],
        "summary": "Create a comment.",
        "description": "User creates a comment.",
        "operationId": "createOneComment",
        "parameters": [
          {
            "name": "authorization",
            "in": "header",
            "description": "User token.",
            "schema": {
              "$ref": "#/components/schemas/bearerToken"
            },
            "required": true
          }
        ],
        "requestBody": {
          "description": "Comment properties.",
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/commentPost"
              }
            }
          },
          "required": true
        },
        "responses": {
          "201": {
            "description": "Comment created."
          },
          "400": {
            "description": "Bad request."
          },
          "401": {
            "description": "Unauthorized request."
          },
          "500": {
            "description": "Server error."
          }
        }
      },
      "get": {
        "tags": ["comments"],
        "summary": "Get all comments.",
        "description": "User gets all comments.",
        "operationId": "getAllComments",
        "parameters": [
          {
            "name": "authorization",
            "in": "header",
            "description": "User token.",
            "schema": {
              "$ref": "#/components/schemas/bearerToken"
            },
            "required": true
          },
          {
            "name": "minDate",
            "in": "query",
            "description": "Get all comments who have been created after this date. Only the ISO 8601 format (YYYY-MM-DDTHH:mm:ss.sssZ) is supported.",
            "schema": {
              "type": "string",
              "example": "2022-13-01T14:48:02.809Z"
            }
          },
          {
            "name": "maxDate",
            "in": "query",
            "description": "Get all comments who have been created before this date. Only the ISO 8601 format (YYYY-MM-DDTHH:mm:ss.sssZ) is supported.",
            "schema": {
              "type": "string",
              "example": "2022-13-01T14:48:02.809Z"
            }
          },
          {
            "name": "limit",
            "in": "query",
            "description": "Get a limited amount of comments.",
            "schema": {
              "type": "number",
              "example": 6
            }
          },
          {
            "name": "sort",
            "in": "query",
            "description": "Returns elements from oldest to newest ('asc') or from newest to oldest ('desc').",
            "schema": {
              "type": "string",
              "example": "desc"
            }
          },
          {
            "name": "fromUserId",
            "in": "query",
            "description": "Get all comments from a specific user.",
            "schema": {
              "type": "string",
              "example": "632dc722c75414645a459eb2"
            }
          },
          {
            "name": "fromPostId",
            "in": "query",
            "description": "Get all comments from a specific post.",
            "schema": {
              "type": "string",
              "example": "632efd854148fdc69cd7e652"
            }
          },
          {
            "name": "reactions",
            "in": "query",
            "description": "Get reactions of the comments.",
            "schema": {
              "type": "boolean",
              "example": true
            }
          }
        ],
        "responses": {
          "200": {
            "description": "Ok.",
            "content": {
              "application/json": {
                "schema": {
                  "type": "array",
                  "items":{
                    "$ref": "#/components/schemas/comment"
                  }
                }
              }
            }
          },
          "400": {
            "description": "Bad request."
          },
          "401": {
            "description": "Unauthorized request."
          },
          "500": {
            "description": "Server error."
          }
        }
      },
      "delete": {
        "tags": ["comments"],
        "summary": "Delete all comments.",
        "description": "Admin deletes all comments.",
        "operationId": "deleteAllComments",
        "parameters": [
          {
            "name": "authorization",
            "in": "header",
            "description": "User token. Must be an admin one.",
            "schema": {
              "$ref": "#/components/schemas/bearerToken"
            },
            "required": true
          }
        ],
        "responses": {
          "200": {
            "description": "Ok."
          },
          "400": {
            "description": "Bad request."
          },
          "403": {
            "description": "Forbidden request."
          },
          "500": {
            "description": "Server error."
          }
        }
      }
    },
    "/comments/{commentId}": {
      "get": {
        "tags": ["comments"],
        "summary": "Get one comment.",
        "description": "User gets properties of a comment.",
        "operationId": "getOneComment",
        "parameters": [
          {
            "name": "authorization",
            "in": "header",
            "description": "User token.",
            "schema": {
              "$ref": "#/components/schemas/bearerToken"
            },
            "required": true
          },
          {
            "name": "commentId",
            "in": "path",
            "description": "Id of comment to return.",
            "schema": {
              "type": "string",
              "example": "632efdbb4148fdc69cd7e654"
            },
            "required": true
          },
          {
            "name": "reactions",
            "in": "query",
            "description": "Get reactions of the comment.",
            "schema": {
              "type": "boolean",
              "example": true
            }
          }
        ],
        "responses": {
          "200": {
            "description": "Ok.",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/comment"
                }
              }
            }
          },
          "400": {
            "description": "Bad request."
          },
          "401": {
            "description": "Unauthorized request."
          },
          "500": {
            "description": "Server error."
          }
        }
      },
      "put": {
        "tags": ["comments"],
        "summary": "Modify one comment.",
        "description": "User modifies a comment of his own, or admin modifies a comment.",
        "operationId": "modifyOneComment",
        "parameters": [
          {
            "name": "authorization",
            "in": "header",
            "description": "User token.",
            "schema": {
              "$ref": "#/components/schemas/bearerToken"
            },
            "required": true
          },
          {
            "name": "commentId",
            "in": "path",
            "description": "Id of comment to modify.",
            "schema": {
              "type": "string",
              "example": "632efdbb4148fdc69cd7e654"
            },
            "required": true
          }
        ],
        "requestBody": {
          "description": "New properties of the comment.",
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/commentPut"
              }
            }
          },
          "required": true
        },
        "responses": {
          "200": {
            "description": "Ok.",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/comment"
                }
              }
            }
          },
          "400": {
            "description": "Bad request."
          },
          "401": {
            "description": "Unauthorized request."
          },
          "500": {
            "description": "Server error."
          }
        }
      },
      "delete": {
        "tags": ["comments"],
        "summary": "Delete one comment.",
        "description": "User deletes a comment of his own, or admin deletes a comment.",
        "operationId": "deleteOneComment",
        "parameters": [
          {
            "name": "authorization",
            "in": "header",
            "description": "User token.",
            "schema": {
              "$ref": "#/components/schemas/bearerToken"
            },
            "required": true
          },
          {
            "name": "commentId",
            "in": "path",
            "description": "Id of comment to delete.",
            "schema": {
              "type": "string",
              "example": "632efdbb4148fdc69cd7e654"
            },
            "required": true
          }
        ],
        "responses": {
          "200": {
            "description": "Ok."
          },
          "400": {
            "description": "Bad request."
          },
          "401": {
            "description": "Unauthorized request."
          },
          "500": {
            "description": "Server error."
          }
        }
      }
    },
...
  },
  "components": {
    "schemas": {
...
      "comment": {
        "title": "Comment",
        "description": "Comment schema.",
        "type": "object",
        "properties": {
          "_id": {
            "type": "string",
            "example": "632efdbb4148fdc69cd7e654"
          },
          "content": {
            "type": "string",
            "example": "Bravo !"
          },
          "postId": {
            "type": "string",
            "example": "632efd854148fdc69cd7e652"
          },
          "userId": {
            "type": "string",
            "example": "632dc722c75414645a459eb2"
          },
          "creationDate": {
            "type": "string",
            "example": "2022-13-01T14:48:02.809Z"
          },
          "reactions": {
            "type": "array",
            "items": {
              "$ref": "#/components/schemas/reaction"
            }
          },
          "__v": {
            "type": "number",
            "example": 0
          }
        },
        "required": [
          "_id",
          "content",
          "postId",
          "userId",
          "creationDate",
          "__v"
        ]
      },
      "reaction": {
        "title": "Reaction",
        "description": "Reaction schema. Exactly one of 'postId' or 'commentId' must be completed.",
        "type": "object",
        "properties": {
          "_id": {
            "type": "string",
            "example": "632efdbb4148fdc69cd7e784"
          },
          "type": {
            "type": "string",
            "example": "Like"
          },
          "postId": {
            "type": "string",
            "example": ""
          },
          "commentId": {
            "type": "string",
            "example": "632efdbb4148fdc69cd7e654"
          },
          "userId": {
            "type": "string",
            "example": "632dc722c75414645a459eb2"
          },
          "creationDate": {
            "type": "string",
            "example": "2022-13-01T14:48:02.809Z"
          },
          "__v": {
            "type": "number",
            "example": 0
          }
        },
        "required": [
          "_id",
          "type",
          "userId",
          "creationDate",
          "__v"
        ]
      },
...
      "bearerToken": {
        "description": "Bearer token schema.",
        "type": "string",
        "example": "Bearer eyJhbGciOiJIUzI1NiIsInR5cCI6IkpXVCJ9.eyJ1c2VySWQiOiI2MzMxZDk1MWU3N2YxNDAxMmQxMjExNmEiLCJpc0FkbWluIjpmYWxzZSwiaWF0IjoxNjY0MjExMzEyLCJleHAiOjE2NjQyOTc3MTJ9.03h-xFcgHmwpYCHMZq36hnSAsHNS7TItboYmAliKSIk"
      },
...
      "commentPost": {
        "title": "CommentPost",
        "description": "Comment schema used when user publishes his comment.",
        "type": "object",
        "properties": {
          "content": {
            "type": "string",
            "example": "Bravo !"
          },
          "postId": {
            "type": "string",
            "example": "632efd854148fdc69cd7e652"
          }
        },
        "required": [
          "content",
          "postId"
        ]
      },
      "commentPut": {
        "title": "CommentPut",
        "description": "Post schema used when user modifies his comment.",
        "type": "object",
        "properties": {
          "content": {
            "type": "string",
            "example": "Félicitations !"
          }
        },
        "required": [
          "content"
        ]
      },
...
    }
  }
}